// active/src/Types/arquivo.ts
export type Photo = {
    name: string;
    url: string;
}

export type data = {
    Tecnico: string;
    Doenca: string;
    Sexo: string;
    Data: string;
    FaixaEtaria: string
    Qtde: number;

}

export type apiE = {
    Tecnico: string;
    Doenca: string;
    Sexo: string;
    Data: string;
    FaixaEtaria: string;
    Qtde: number;
    r: string;
    text: string;
}